Initial draft of UML-style state machines (statemachine.py).
This is the first attempt at cleaning up the old FSM code. Instead
of subclassing the base State class (which was strange, since we only
ever needed a single instance of each state), we now treat each state
as an instance/object of the class. This drastically changes how we
must write the individual state's functionality; that still needs to
be documented and tested.